+++ src/blockchain/service.rs
@@ -24,7 +24,7 @@ use crate::{
     channel::confirmed_channel,
     db::Db,
     wm::Wm,
-    Transaction,
+    KeyPair, Transaction,
 };
 use std::sync::Arc;
 use std::thread::{self, JoinHandle};
@@ -38,6 +38,8 @@ pub struct BlockConfig {
     pub timeout: u16,
     /// Blockchain network identifier.
     pub network: String,
+    /// Node KeyPair
+    pub keypair: Arc<KeyPair>,
 }
 
 /// Block service data.
@@ -182,6 +184,7 @@ mod tests {
             threshold: 42,
             timeout: 3,
             network: "skynet".to_string(),
+            keypair: Arc::new(crate::crypto::sign::tests::create_test_keypair()),
         };
 
         let is_validator = is_validator_function();